Poll-wait for server power state change in PowerUp/PowerDown apis (#33)
* Fix flake8 errors in serveractions.py

Three lines in serveractions.py which were longer than 79
chars and caused flake8 to fail.  Split those lines so
flake8 passes now.

* Add polling to power up/down functions

This patch adds logic to the power up/down functions in the apis
folder to poll the server and return only after the desired power
state has happened been reached.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Refactor server polling

The server polling loop is identical between functions except
for the state its watching for.  So refactored the loop into
its own function that takes in a state.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add polling to graceful power down and power cycle

Added the polling loop to both the graceful power
down and the power cycle functions.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Address first round of commends in PR

Changed the error raised to be a ImcOperationError.

Signed-off-by: Britt Houser <bhouser@cisco.com>