Vmedia updates and unit-tests (#38)
* Per best practices, move all imports to top of module

Python best practices [0] say todo all your imports at the top
of a given module. This commit moves all imports in the
apis/servers/remotepresence module to the top.

[0] https://docs.python.org/2/faq/programming.html

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Remove device_type from vmedia_mount_add

When device_type is set when adding a vmedia mount, an
error is thrown stating that is not a writable field.
Removing setting of device_type.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add function to remove all virtual media

If a server is not cleaned up properly between
uses, there can exist leftover virtual media
mounts.  The current user has no knowledge of
these mounts, and so they cannot delete them by
volume name.  So this function is added to
allow a user to clean out all mounts to have
a clean slate before adding new mounts.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Add function to mount ISO from URL

A common use case for vmedia is to mount an ISO from a URI.
This commit adds a  handy to have a function that a user
can pass a URI, and the function breaks it down and
dynamically determine the protocol,etc and makes the
correct call to vmedia_mount_add.

Signed-off-by: Britt Houser <bhouser@cisco.com>

* Create unit_tests folder

This commit creates a tests/unit_tests/ folder for containing
mock based unit tests which do not require any physical hardware.